CORRECT INTEGRATION TESTS CONFIGURATION
after each method deleted tables, not rows
Delete database provided it is test database

Roldán Cadena, Jesús <@jesrolcad>

+++ app/tests/integration/conftest.py
@@ -1,8 +1,10 @@
 import os
 import pytest
 from sqlalchemy_utils import database_exists, create_database, drop_database
+
 os.environ["ENVIRONMENT"] = "TEST"
-from app.db.database import Base, engine
+
+from app.db.database import Base, engine, get_db_as_context_manager
 
 
 def import_db_entities():
@@ -17,16 +19,16 @@ def before_all():
         import_db_entities()
         Base.metadata.create_all(engine)
 
-
-
 @pytest.fixture(scope="session", autouse=True)
 def after_all():
     yield
-    drop_database(engine.url)
-
-
+    print("ENGINE URL DATABASE: ", engine.url.database)
+    if(engine.url.database == os.getenv('TEST_POSTGRES_DB')):
+        drop_database(engine.url)
 
 @pytest.fixture(scope="function", autouse=True)
 def after_each():
     yield
-    Base.metadata.drop_all(engine)
+    with get_db_as_context_manager() as session:
+        for table in reversed(Base.metadata.sorted_tables):
+            session.execute(table.delete())
